tests/wpt/mozilla/tests/mozilla/promise.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  promise_test(function() {
      var t = new TestBinding;
      return t.returnResolvedPromise('success')
          .then(function(s) {
              assert_equals(s, 'success');
          });
  }, 'Resolve callback gets argument');

  promise_test(function(test) {
      var t = new TestBinding;
      return t.returnRejectedPromise('success')
          .then(test.unreached_func())
          .catch(function(s) {
              assert_equals(s, 'success');
          });
  }, 'Reject callback gets argument');

  promise_test(function(test) {
      var t = new TestBinding;
      var resolved;
      var p = t.promiseNativeHandler(function(v) {
          resolved = v;
      }, null);
      t.promiseResolveNative(p, 'success');
      return p.then(function() {
          assert_equals(resolved, 'success')
      });
  }, 'Native resolve callback gets argument');

  promise_test(function(test) {
      var t = new TestBinding;
      var rejected;
      var p = t.promiseNativeHandler(null, function(v) {
          rejected = v;
      });
      t.promiseRejectNative(p, 'success');
      return p.catch(function() {
          assert_equals(rejected, 'success')
      });
  }, 'Native reject callback gets argument');
</script>
